% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_exr.R
\name{get_historical_exr}
\alias{get_historical_exr}
\title{Retrieve exchange rates over time from frankfurter.app}
\usage{
get_historical_exr(
  from_currency,
  to_currency,
  start_date,
  end_date = Sys.Date()
)
}
\arguments{
\item{from_currency}{currency you want to convert from}

\item{to_currency}{currency you want to convert to}

\item{start_date}{start date of time-series}

\item{end_date}{end date of time-series}
}
\description{
Retrieves a time series of a exchange rate between specified dates
}
