% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_exr.R
\name{get_latest_exr}
\alias{get_latest_exr}
\title{Retrieve latest exchange rate frankfurter.app}
\usage{
get_latest_exr(from_currency, to_currency)
}
\arguments{
\item{from_currency}{currency you want to convert from}

\item{to_currency}{currency you want to convert to}
}
\description{
Retrieves the latest exchange rate between specified dates
}
